Refactor handlers to different files
